add support for branches other than master - fixes #2
removing the hardcoded reference to the master branch and replacing
it with a regex match on the .git/HEAD file.  This should match on
a branch or a commit.  If neither of those are matched for any
reason, the branch will default back to master and an issue should
be opened outlining the contents of the HEAD file.